+++ HISTORY.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## x.x.x (not yet released)
+* Add additional protected methods to JsonCodec to allow subclasses to
+  further customize the encoding used.
 
 ## 2.1.0 (2023-10-02)
 * Allow ::jsonClassHintFor() to return a class-string suffixed with